requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.7.4
cloudinary==1.40.0
distro==1.9.0
Django==4.2.14
django-otp==1.5.1
djangorestframework==3.15.2
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
mysqlclient==2.2.4
openai==1.37.0
pillow==10.4.0
pkg_resources==0.0.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
